+++ dbupgrade/__init__.py
@@ -1,3 +1,7 @@
+from .result import (  # noqa: F401
+    UpgradeResult as UpgradeResult,
+    VersionResult as VersionResult,
+)
 from .upgrade import db_upgrade as db_upgrade  # noqa: F401
 from .version import (  # noqa: F401
     MAX_API_LEVEL as MAX_API_LEVEL,