+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# Unreleased Changes
+
 # News in dbupgrade 2023.10.0
 
 Drop support for Python 3.7 and 3.8.