+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased Changes
 
+Add `UpgradeResult.was_upgraded` property.
+
 # News in dbupgrade 2023.10.0
 
 Drop support for Python 3.7 and 3.8.