+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Add `UpgradeResult.was_upgraded` property.
 
+Re-export `UpgradeResult` and `VersionResult` from `dbupgrade`.
+
 # News in dbupgrade 2023.10.0
 
 Drop support for Python 3.7 and 3.8.